# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: AI Employee SaaS API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /assistants/{assistantId}/leads:
    get:
      operationId: listLeads
      tags: [leads]
      summary: List leads captured by an assistant
      security:
        - clerkAuth: []
      parameters:
        - name: assistantId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of leads
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Lead"
    post:
      operationId: createLead
      tags: [leads]
      summary: Capture a new lead
      parameters:
        - name: assistantId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLeadBody"
      responses:
        "201":
          description: Created lead
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
  /leads/{id}:
    patch:
      operationId: updateLead
      tags: [leads]
      summary: Update a lead status
      security:
        - clerkAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateLeadBody"
      responses:
        "200":
          description: Updated lead
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"

  # ── Appointments ─────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    clerkAuth:
      type: http
      scheme: bearer
  schemas:
...
    Lead:
      type: object
      properties:
        id:
          type: integer
        assistantId:
          type: integer
        conversationId:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        status:
          type: string
          enum: [new, contacted, qualified, converted, lost]
        notes:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, assistantId, name, status, createdAt]
    CreateLeadBody:
      type: object
      properties:
        assistantId:
          type: integer
        conversationId:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: string
      required: [assistantId, name]
    UpdateLeadBody:
      type: object
      properties:
        status:
          type: string
          enum: [new, contacted, qualified, converted, lost]
        notes:
          type: string
        name:
          type: string
        email:
          type: string
        phone:
          type: string
...